backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
influxdb-client==1.38.0

